Refatorando projeto para EJS
