+++ app/public/css/forms.css
@@ -1,0 +1,74 @@
+/*
+Sass::SyntaxError: An @import loop has been found: D:/aimbra/login-system/public/assets/sass/forms.scss imports itself
+
+Backtrace:
+D:/aimbra/login-system/public/assets/sass/forms.scss:
+D:/aimbra/login-system/public/assets/sass/forms.scss:1
+D:/aimbra/login-system/public/assets/sass/forms.scss:1
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:435:in `handle_import_loop!'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:219:in `visit_import'
+org/jruby/RubyKernel.java:2096:in `send'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/base.rb:37:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:98:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:224:in `visit_import'
+org/jruby/RubyArray.java:2336:in `collect'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:224:in `visit_import'
+org/jruby/RubyKernel.java:2096:in `send'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/base.rb:37:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:98:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/base.rb:53:in `visit_children'
+org/jruby/RubyArray.java:2336:in `collect'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/base.rb:53:in `visit_children'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:107:in `visit_children'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:119:in `with_environment'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:106:in `visit_children'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/base.rb:37:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:126:in `visit_root'
+org/jruby/RubyKernel.java:2096:in `send'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/base.rb:37:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:98:in `visit'
+org/jruby/RubyKernel.java:2096:in `send'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/visitors/perform.rb:7:in `visit'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/tree/root_node.rb:20:in `render'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/engine.rb:315:in `_render'
+D:/Scout-Installer/vendor/gems/gems/sass-3.2.1/lib/sass/../sass/engine.rb:262:in `render'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:140:in `compile'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:126:in `timed'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:139:in `compile'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/logger.rb:45:in `red'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:138:in `compile'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:118:in `compile_if_required'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:103:in `run'
+org/jruby/RubyArray.java:1603:in `each'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:101:in `run'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:126:in `timed'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/compiler.rb:100:in `run'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/commands/watch_project.rb:147:in `recompile'
+org/jruby/RubyMethod.java:133:in `call'
+org/jruby/RubyProc.java:268:in `call'
+org/jruby/RubyProc.java:228:in `call'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/path.rb:73:in `run_callback'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/path.rb:55:in `callback_action'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/path.rb:35:in `update'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/state/directory.rb:39:in `modified'
+org/jruby/RubyArray.java:1603:in `each'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/state/directory.rb:37:in `modified'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/state/directory.rb:18:in `refresh'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/backends/polling.rb:17:in `run'
+org/jruby/RubyArray.java:1603:in `each'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/backends/polling.rb:17:in `run'
+org/jruby/RubyKernel.java:1419:in `loop'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/backends/polling.rb:15:in `run'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm/monitor.rb:26:in `run'
+D:/Scout-Installer/vendor/gems/gems/fssm-0.2.9/lib/fssm.rb:70:in `monitor'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/commands/watch_project.rb:87:in `perform'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/commands/base.rb:18:in `execute'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/commands/project_base.rb:19:in `execute'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/exec/sub_command_ui.rb:43:in `perform!'
+D:/Scout-Installer/vendor/gems/gems/compass-0.12.2/lib/compass/exec/sub_command_ui.rb:15:in `run!'
+D:\Scout-Installer\bin\compass:12:in `(root)'
+*/
+body:before {
+  white-space: pre;
+  font-family: monospace;
+  content: "Sass::SyntaxError: An @import loop has been found: D:/aimbra/login-system/public/assets/sass/forms.scss imports itself"; }
