+++ app/public/css/style.css
@@ -1,0 +1,252 @@
+@import url("https://use.typekit.net/sys7rxc.css");
+/* line 3, ../sass/_navbar.scss */
+.navbar {
+  background-color: #00233d !important;
+  padding: 20px 0;
+  color: #fff;
+}
+/* line 11, ../sass/_navbar.scss */
+.navbar .navbar-brand img {
+  width: 80%;
+  height: auto;
+  -webkit-filter: brightness(1.2);
+  -moz-filter: brightness(1.2);
+  -ms-filter: brightness(1.2);
+  -o-filter: brightness(1.2);
+  filter: brightness(1.2);
+}
+/* line 19, ../sass/_navbar.scss */
+.navbar .navbar-brand img:hover {
+  -webkit-filter: brightness(1.5);
+  -moz-filter: brightness(1.5);
+  -ms-filter: brightness(1.5);
+  -o-filter: brightness(1.5);
+  filter: brightness(1.5);
+}
+/* line 31, ../sass/_navbar.scss */
+.navbar li.nav-item a.nav-link {
+  color: #fff;
+}
+/* line 33, ../sass/_navbar.scss */
+.navbar li.nav-item a.nav-link:hover {
+  color: #00d48d;
+}
+/* line 40, ../sass/_navbar.scss */
+.navbar li.active {
+  color: #00d48d;
+}
+/* line 42, ../sass/_navbar.scss */
+.navbar li.active a {
+  color: #00d48d;
+}
+
+/* line 3, ../sass/_forms.scss */
+.btn {
+  padding: 1rem 2rem;
+}
+
+/* line 7, ../sass/_forms.scss */
+.btn-success {
+  border-width: 0;
+  background-color: #00d48d;
+}
+/* line 15, ../sass/_forms.scss */
+.btn-success:hover {
+  background-color: #01ffaa;
+}
+
+/* line 22, ../sass/_forms.scss */
+.btn-outline-success {
+  border-color: #00d48d;
+  color: #00d48d;
+}
+/* line 28, ../sass/_forms.scss */
+.btn-outline-success:hover {
+  background-color: #00d48d;
+}
+
+/* line 38, ../sass/_forms.scss */
+.custom-select.is-valid,
+.form-control.is-valid,
+.was-validated .custom-select:valid,
+.was-validated .form-control:valid {
+  border-color: #00d48d;
+}
+
+/* line 44, ../sass/_forms.scss */
+.valid-feedback {
+  color: #00d48d;
+}
+
+/* line 48, ../sass/_forms.scss */
+textarea.form-control {
+  height: 150px;
+}
+
+/* line 52, ../sass/_forms.scss */
+select.custom-select {
+  height: 55px;
+}
+
+/* line 56, ../sass/_forms.scss */
+input.form-control, textarea.form-control {
+  padding: 1rem;
+}
+
+/* line 60, ../sass/_forms.scss */
+.form-access {
+  margin-top: 5rem;
+}
+/* line 66, ../sass/_forms.scss */
+.form-access p {
+  text-align: center;
+  margin: 0 auto;
+  width: 30%;
+}
+/* line 75, ../sass/_forms.scss */
+.form-access .form-area form {
+  padding-right: 30%;
+  padding-left: 30%;
+}
+/* line 80, ../sass/_forms.scss */
+.form-access .form-area form label {
+  text-align: left;
+}
+
+/* line 4, ../sass/_fonts.scss */
+body {
+  font-family: proxima-nova, sans-serif;
+  font-style: normal;
+  font-weight: 100;
+}
+
+/* line 12, ../sass/_fonts.scss */
+h1, h2, h3, h4, h5, h6 {
+  font-weight: 700;
+  color: #00233d !important;
+}
+
+/* line 19, ../sass/_fonts.scss */
+.title-decorator-center {
+  display: inline-block;
+  position: relative;
+  text-decoration: none;
+}
+/* line 26, ../sass/_fonts.scss */
+.title-decorator-center::before, .title-decorator-center::after {
+  content: "";
+  position: absolute;
+  top: 57%;
+  width: 60%;
+  height: 1px;
+  background-color: gainsboro;
+}
+/* line 34, ../sass/_fonts.scss */
+.title-decorator-center::before {
+  left: -85%;
+}
+/* line 37, ../sass/_fonts.scss */
+.title-decorator-center::after {
+  right: -85%;
+}
+
+/* line 42, ../sass/_fonts.scss */
+p {
+  font-size: 1.2rem;
+}
+
+/* line 48, ../sass/_fonts.scss */
+a {
+  color: #00d48d;
+}
+/* line 50, ../sass/_fonts.scss */
+a:hover {
+  color: #01ffaa;
+}
+
+/* line 7, ../sass/_template.scss */
+footer {
+  width: 100%;
+  position: absolute;
+  bottom: 0;
+}
+
+/* line 14, ../sass/_template.scss */
+header .dropdown-menu {
+  min-width: 30rem;
+  padding: .5rem 10px;
+}
+/* line 18, ../sass/_template.scss */
+header .dropdown-item {
+  padding: 1.25rem 1.5rem;
+}
+/* line 20, ../sass/_template.scss */
+header .dropdown-item:hover {
+  border-radius: 5px;
+  background-color: #00d48d;
+  color: #fff;
+  -webkit-animation: fadein 2s;
+  /* Safari, Chrome and Opera > 12.1 */
+  -moz-animation: fadein 2s;
+  /* Firefox < 16 */
+  -ms-animation: fadein 2s;
+  /* Internet Explorer */
+  -o-animation: fadein 2s;
+  /* Opera < 12.1 */
+  animation: fadein 2s;
+}
+
+/* line 1, ../sass/_alerts.scss */
+.message {
+  position: absolute;
+  right: 2rem;
+  padding: 1rem 2rem;
+  font-size: 0.8rem;
+  font-weight: 500;
+}
+
+/* line 1, ../sass/_icons.scss */
+.fab {
+  font-size: 1.4rem;
+}
+
+/* line 7, ../sass/_icons.scss */
+.fa-facebook-square {
+  color: #0038B9;
+}
+
+/* line 10, ../sass/_icons.scss */
+.fa-twitter-square {
+  color: #149EE6;
+}
+
+/* line 9, ../sass/style.scss */
+* {
+  margin: 0;
+  padding: 0;
+}
+
+/* line 14, ../sass/style.scss */
+main.home {
+  margin-top: 100px;
+}
+/* line 21, ../sass/style.scss */
+main.home figure img {
+  width: 10%;
+}
+/* line 26, ../sass/style.scss */
+main.home h1 {
+  font-weight: 900;
+  text-transform: uppercase;
+}
+/* line 36, ../sass/style.scss */
+main.home .social-group ul {
+  margin-left: 0;
+  margin-top: -10px;
+  padding: 0;
+}
+/* line 42, ../sass/style.scss */
+main.home .social-group ul li {
+  display: inline-block;
+  margin-right: 10px;
+}
