+++ app/public/js/script.js
@@ -1,0 +1,5 @@
+function login() {
+    let message = document.querySelector('.message')
+    message.removeAttribute('hidden')
+    message.innerHTML = 'Invalid password or username'
+}
