+++ app/public/sass/_alerts.scss
@@ -1,0 +1,9 @@
+.message {
+    position: absolute;
+    right: 2rem;
+    padding: 1rem 2rem;
+    font: {
+        size: 0.8rem;
+        weight: 500;
+    }
+}
