+++ app/public/sass/_colors.scss
@@ -1,0 +1,7 @@
+$gray-100: rgb(240, 240, 240);
+$gray-200: rgb(220, 220, 220);
+
+$primary: #00D48D !default;
+$primary-hover: rgb(1, 255, 170) !default;
+$secondary: #0093D4 !default;
+$blue-dark: #00233D !important;
