+++ app/public/sass/_icons.scss
@@ -1,0 +1,12 @@
+.fab {
+    font: {
+        size: 1.4rem;
+    }
+}
+
+.fa-facebook-square {
+    color: #0038B9;
+}
+.fa-twitter-square {
+    color: #149EE6;
+}
