+++ app/routes/about.js
@@ -1,0 +1,5 @@
+module.exports = (app) =>
+    app.get(
+        '/about',
+        (req, resp) => resp.render('about/about')
+    )
