+++ app/routes/contact.js
@@ -1,0 +1,5 @@
+module.exports = (app) =>
+    app.get(
+        '/contact',
+        (req, resp) => resp.render('contact/contact')
+    )
