+++ app/routes/home.js
@@ -1,0 +1,5 @@
+module.exports = (app) =>
+    app.get(
+        '/',
+        (req, resp) => resp.render('home/index')
+    )
