+++ app/routes/signin.js
@@ -1,0 +1,5 @@
+module.exports = (app) =>
+    app.get(
+        '/signin',
+        (req, resp) => resp.render('security/signin')
+    )
