+++ app/routes/signup.js
@@ -1,0 +1,5 @@
+module.exports = (app) =>
+    app.get(
+        '/signup',
+        (req, resp) => resp.render('security/signup')
+    )
