+++ app/views/about/about.ejs
@@ -1,0 +1,2 @@
+<% include ../includes/header.ejs %>
+<% include ../includes/footer.ejs %>
