+++ app/views/contact/contact.ejs
@@ -1,0 +1,49 @@
+<% include ../includes/header.ejs %>
+    <main>
+        <section class="container form-access">
+            <div class="text-center">
+                <div class="message alert alert-danger animated bounceInRight" hidden></div>
+
+                <h1 class="title-decorator-center">Contact with me</h1>
+                <p class="mt-3">Please, you need to fill in the fields below. We will respond as soon as possible.</p>
+
+                <div class="form-area">
+                    <!-- http://localhost:3003/usuarios -->
+                    <form class="mt-5" novalidate>
+                        <div class="form-group">
+                            <label for="name" class="text-left" hidden>Name</label>
+                            <input type="text" class="form-control" id="name" name="name" aria-describedby="name" placeholder="Enter your name">
+                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                        </div>
+                        <div class="form-group">
+                            <label for="email" class="text-left" hidden>Email address</label>
+                            <input type="email" class="form-control" id="email" name="email" aria-describedby="emailHelp" placeholder="Enter email">
+                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                        </div>
+                        <div class="form-group">
+                            <label for="email" class="text-left" hidden>Email address</label>
+                            <select class="custom-select" name="subject" id="subject">
+                                <option>Select one subject</option>
+                                <option value="budget">budget request</option>
+                                <option value="suport">Suport</option>
+                                <option value="outhers">Outhers</option>
+                            </select>
+                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                        </div>
+                        <div class="form-group">
+                            <textarea class="form-control"></textarea>
+                        </div>
+                        <div class="mt-5 form-inline">
+                            <div class="custom-control custom-checkbox mr-auto">
+                            </div>
+                            <button class="btn btn-success mr-0 register" type="click" onclick="sendContact()">Send Message</button>
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </section>
+    </main>
+<% include ../includes/footer.ejs %>
