+++ app/views/home/index.ejs
@@ -1,0 +1,32 @@
+<% include ../includes/header.ejs %>
+<main class="home">
+    <section class="container">
+        <div class="col-xs-12">
+            <figure class="text-center">
+                <img src="img/brand.svg" alt="Aimbra brand from logotype">
+            </figure>
+            <div class="content">
+                <h1 class="underline text-center">Under
+                    <br>Constructor</h1>
+                <h5 class="text-center">Sorry, We are under constructor</h5>
+                <hr class="mt-5">
+                <div class="social-group text-center">
+                    <p class="text-center">Flow on Social Networks:</p>
+                    <ul class="text-center">
+                        <li>
+                            <a href="#">
+                                <i class="fab fa-facebook-square"></i>
+                            </a>
+                        </li>
+                        <li>
+                            <a href="#">
+                                <i class="fab fa-twitter-square"></i>
+                            </a>
+                        </li>
+                    </ul>
+                </div>
+            </div>
+        </div>
+    </section>
+</main>
+<% include ../includes/footer.ejs %>
