+++ app/views/security/signin.ejs
@@ -1,0 +1,43 @@
+<% include ../includes/header.ejs %>
+    <main>
+        <section class="container form-access">
+            <div class="text-center">
+                <div class="message alert alert-danger animated bounceInRight" hidden>Preencha os dados que estão faltando, por gentileza seu merda!</div>
+                <figure>
+                    <img src="img/brand.png" alt="">
+                </figure>
+                <h1 class="title-decorator-center">System Aimbra Login</h1>
+                <p class="mt-3">Please, you need enter your datas for administrator structores.</p>
+
+                <div class="form-area">
+                    <form action="" class="mt-5" novalidate>
+                        <div class="form-group">
+                            <label for="email" class="text-left" hidden>Email address</label>
+                            <input type="text" class="form-control" id="email" name="email" aria-describedby="emailHelp" placeholder="Enter email or username">
+                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                        </div>
+                        <div class="form-group">
+                            <label for="password" class="text-left" hidden>Password</label>
+                            <input type="password" class="form-control" id="password" aria-describedby="emailHelp" placeholder="Enter your password">
+                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                        </div>
+                        <div class="mt-5 form-inline">
+                            <h6 class="mr-auto mr-1">
+                                <a href="#">Forgot password?</a>
+                            </h6>
+                            <button class="btn btn-success mr-0 btn-login" onclick="login()" type="button">Login</button>
+                        </div>
+                        <div class="mt-5">
+                            <hr>
+                            <h6 class="text-center">
+                                <a href="/signup.html">Not account? Register Now!</a>
+                            </h6>
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </section>
+    </main>
+<% include ../includes/footer.ejs %>
