+++ app/views/security/signup.ejs
@@ -1,0 +1,65 @@
+<% include ../includes/header.ejs %>
+
+<main>
+    <section class="container form-access">
+        <div class="text-center">
+            <div class="message alert alert-danger animated bounceInRight">Sorry! Sign up area is disable by one time...</div>
+            <figure>
+                <img src="img/brand.png" alt="">
+            </figure>
+            <h1 class="title-decorator-center">Register</h1>
+            <p class="mt-3">Please, you need enter your datas for administrator structores.</p>
+
+            <div class="form-area">
+                <!-- http://localhost:3003/usuarios -->
+                <form class="mt-5" novalidate>
+                    <div class="form-group">
+                        <label for="username" class="text-left" hidden>Username</label>
+                        <input type="text" class="form-control" id="username" name="username" aria-describedby="username" placeholder="Enter username">
+                        <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                        <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                    </div>
+                    <div class="form-group">
+                        <label for="email" class="text-left" hidden>Email address</label>
+                        <input type="email" class="form-control" id="email" name="email" aria-describedby="emailHelp" placeholder="Enter email">
+                        <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                        <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                    </div>
+                    <div class="form-group">
+                        <label for="password" class="text-left" hidden>Email address</label>
+                        <input type="password" class="form-control" id="password" name="password" aria-describedby="emailHelp" placeholder="Enter Password">
+                        <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                        <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                    </div>
+                    <div class="form-group">
+                        <label for="repeatPassword" class="text-left" hidden>Repeat Password</label>
+                        <input type="repeatPassword" class="form-control" name="repeatPassword" id="repeatPassword" aria-describedby="emailHelp"
+                            placeholder="Repeat your password">
+                        <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
+                        <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
+                    </div>
+                    <div class="mt-5 form-inline">
+                        <div class="custom-control custom-checkbox mr-auto">
+                            <input type="checkbox" class="custom-control-input" name="termsAndConditions" id="termsAndConditions" required>
+                            <label class="custom-control-label" for="termsAndConditions">I accept
+                                <em>
+                                    <a href="#"> Terms and Conditions</a>
+                                </em>
+                            </label>
+                            <div class="invalid-feedback">You need seleted Terms and Conditions.</div>
+                        </div>
+                        <button class="btn btn-success mr-0 register" type="click" onclick="registerValidate()">Register</button>
+                    </div>
+                    <div class="mt-5">
+                        <hr>
+                        <h6 class="text-center">
+                            <a href="/signin.html">Do you have account system? Sign in!</a>
+                        </h6>
+                    </div>
+                </form>
+            </div>
+        </div>
+    </section>
+</main>
+
+<% include ../includes/footer.ejs %>
