--- contact.html
@@ -1,121 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Aimbra - Login System</title>
-
-    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/animate.css@3.5.2/animate.min.css">
-    <link rel="stylesheet" href="public/assets/css/sign.css">
-
-
-</head>
-
-<body>
-    <header>
-        <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-            <section class="container">
-                <a class="navbar-brand" href="/index.html">
-                    <picture>
-                        <source srcset="public/assets/img/logo-expand.png" media="(min-width: 1024px)">
-                        <img src="public/assets/img/brand.png" alt="Brand da Aimbra softwares">
-                    </picture>
-                </a>
-
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
-                    aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-
-                <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                    <ul class="navbar-nav ml-auto mr-3">
-                        <!-- <li class="nav-item active"><a class="nav-link" href="about.html">About</a></li>
-                        <li class="nav-item"><a class="nav-link" href="contact.html">Suport</a></li> -->
-                        <li class="nav-item dropdown">
-                            <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true"
-                                aria-expanded="false">
-                                Menu
-                            </a>
-                            <div class="dropdown-menu animated bounceInDown" aria-labelledby="navbarDropdown">
-                                <a class="dropdown-item" href="/about.html">About</a>
-                                <a class="dropdown-item" href="/contact.html">Contact</a>
-                                <div class="dropdown-divider"></div>
-                                <a class="dropdown-item" href="suport.html">Suport</a>
-                            </div>
-                        </li>
-                    </ul>
-                    <div class="form-inline my-2 my-lg-0">
-                        <button class="btn btn-outline-success my-2 my-sm-0 mr-3" onclick="javascript: window.location.href ='/signin.html'">Login</button>
-                        <button class="btn btn-success my-2 my-sm-0" onclick="javascript: window.location.href ='/signup.html'">Register</button>
-                    </div>
-                </div>
-            </section>
-        </nav>
-    </header>
-
-    <main>
-        <section class="container form-access">
-            <div class="text-center">
-                <div class="message alert alert-danger animated bounceInRight" hidden></div>
-
-                <h1 class="title-decorator-center">Contact with me</h1>
-                <p class="mt-3">Please, you need to fill in the fields below. We will respond as soon as possible.</p>
-
-                <div class="form-area">
-                    <!-- http://localhost:3003/usuarios -->
-                    <form class="mt-5" novalidate>
-                        <div class="form-group">
-                            <label for="name" class="text-left" hidden>Name</label>
-                            <input type="text" class="form-control" id="name" name="name" aria-describedby="name" placeholder="Enter your name">
-                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
-                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
-                        </div>
-                        <div class="form-group">
-                            <label for="email" class="text-left" hidden>Email address</label>
-                            <input type="email" class="form-control" id="email" name="email" aria-describedby="emailHelp" placeholder="Enter email">
-                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
-                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
-                        </div>
-                        <div class="form-group">
-                            <label for="email" class="text-left" hidden>Email address</label>
-                            <select class="custom-select" name="subject" id="subject">
-                                <option>Select one subject</option>
-                                <option value="budget">budget request</option>
-                                <option value="suport">Suport</option>
-                                <option value="outhers">Outhers</option>
-                            </select>
-                            <div class="valid-feedback text-left animated fadeIn">Looks good!</div>
-                            <div class="invalid-feedback text-left animated fadeIn">Sorry!</div>
-                        </div>
-                        <div class="form-group">
-                            <textarea class="form-control"></textarea>
-                        </div>
-                        <div class="mt-5 form-inline">
-                            <div class="custom-control custom-checkbox mr-auto">
-                            </div>
-                            <button class="btn btn-success mr-0 register" type="click" onclick="sendContact()">Send Message</button>
-                        </div>
-                    </form>
-                </div>
-            </div>
-        </section>
-    </main>
-
-    <footer>
-        <section class="container text-center mt-5">
-            <hr>
-            <p>
-                Copyright © 2018 Aimbra Systems. All rights reserved.
-            </p>
-        </section>
-    </footer>
-
-    <script src="./../node_modules/jquery/dist/jquery.min.js"></script>
-    <script src="./../node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
-    <script src="./index.js"></script>
-</body>
-
-</html>
