--- index.html
@@ -1,94 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <meta http-equiv="X-UA-Compatible" content="ie=edge">
-    <title>Aimbra - Login System</title>
-
-    <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.0.13/css/all.css" integrity="sha384-DNOHZ68U8hZfKXOrtjWvjxusGo9WQnrNx2sqG0tfsghAvtVlRW3tvkXWZh58N9jp"
-        crossorigin="anonymous">
-
-    <link rel="stylesheet" href="node_modules/bootstrap/dist/css/bootstrap.min.css">
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/animate.css@3.5.2/animate.min.css">
-    <link rel="stylesheet" href="public/assets/css/index.css">
-</head>
-<body>
-    <header>
-        <nav class="navbar navbar-expand-lg navbar-dark bg-dark">
-            <section class="container">
-                <a class="navbar-brand" href="/index.html">
-                    <picture>
-                        <source srcset="public/assets/img/logo-expand.png" media="(min-width: 1024px)">
-                        <img src="public/assets/img/brand.png" alt="Brand da Aimbra softwares">
-                    </picture>
-                </a>
-
-                <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#navbarSupportedContent" aria-controls="navbarSupportedContent"
-                    aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                </button>
-
-                <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                    <ul class="navbar-nav ml-auto mr-3">
-                        <!-- <li class="nav-item active"><a class="nav-link" href="about.html">About</a></li>
-                        <li class="nav-item"><a class="nav-link" href="contact.html">Suport</a></li> -->
-                        <li class="nav-item dropdown">
-                            <a class="nav-link dropdown-toggle" href="#" id="navbarDropdown" role="button" data-toggle="dropdown" aria-haspopup="true"
-                                aria-expanded="false">
-                                Menu
-                            </a>
-                            <div class="dropdown-menu animated bounceInDown" aria-labelledby="navbarDropdown">
-                                <a class="dropdown-item" href="/about.html">About</a>
-                                <a class="dropdown-item" href="/contact.html">Contact</a>
-                                <div class="dropdown-divider"></div>
-                                <a class="dropdown-item" href="suport.html">Suport</a>
-                            </div>
-                        </li>
-                    </ul>
-                    <div class="form-inline my-2 my-lg-0">
-                        <button class="btn btn-outline-success my-2 my-sm-0 mr-3" onclick="javascript: window.location.href ='/signin.html'">Login</button>
-                        <button class="btn btn-success my-2 my-sm-0" onclick="javascript: window.location.href ='/signup.html'">Register</button>
-                    </div>
-                </div>
-            </section>
-        </nav>
-    </header>
-
-    <main class="home">
-        <section class="container">
-            <div class="col-xs-12">
-                <figure class="text-center">
-                    <img src="/public/assets/img/brand.svg" alt="Aimbra brand from logotype">
-                </figure>
-                <div class="content">
-                    <h1 class="underline text-center">Under<br>Constructor</h1>
-                    <h5 class="text-center">Sorry, We are under constructor</h5>
-                    <hr class="mt-5">
-                    <div class="social-group text-center">
-                        <p class="text-center">Flow on Social Networks:</p>
-                        <ul class="text-center">
-                            <li><a href="#"><i class="fab fa-facebook-square"></i></a></li>
-                            <li><a href="#"><i class="fab fa-twitter-square"></i></a></li>
-                        </ul>
-
-                    </div>
-                </div>
-            </div>
-        </section>
-    </main>
-
-    <footer>
-        <section class="container text-center mt-5">
-            <hr>
-            <p>
-                Copyright © 2018 Aimbra Systems. All rights reserved.
-            </p>
-        </section>
-    </footer>
-    <script src="./../node_modules/jquery/dist/jquery.min.js"></script>
-    <script src="./../node_modules/popper.js/dist/popper.min.js"></script>
-    <script src="./../node_modules/bootstrap/dist/js/bootstrap.min.js"></script>
-    <script src="./index.js"></script>
-</body>
-</html>
