+++ package.json
@@ -17,7 +17,10 @@
   "dependencies": {
     "body-parser": "^1.18.3",
     "bootstrap": "^4.1.1",
+    "consign": "^0.1.6",
+    "ejs": "^2.6.1",
     "express": "^4.16.3",
+    "express-validator": "^5.2.0",
     "jquery": "^3.3.1",
     "popper.js": "^1.14.3"
   },
