--- public/assets/sass/sign.scss
@@ -1,11 +1,0 @@
-@import './navbar';
-@import './colors';
-@import './forms';
-@import './fonts';
-@import './template';
-@import './alerts';
-
-* {
-    margin: 0;
-    padding: 0;
-}
