--- server.js
@@ -1,17 +1,0 @@
-const express = require('express')
-const app = express()
-const bodyParser = require('body-parser')
-
-app.use (bodyParser.urlencoded({
-    extended: true
-}))
-
-app.post(
-    '/usuarios',
-    (req, resp) => {
-        console.log(req.body)
-        resp.send('<h1>Usuário cadastrado</h1>')
-    }
-)
-
-app.listen(3003)
